本地音乐增加播放按钮，增加播放状态及动画。

+++ app/src/main/java/cn/ommiao/musicmiao/utils/UIUtil.java
@@ -6,7 +6,7 @@ import cn.ommiao.musicmiao.core.App;
 
 public class UIUtil {
 
-    public static int dp2px(int dpVal) {
+    public static int dp2px(float dpVal) {
         return (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP,
                 dpVal, App.getContext().getResources().getDisplayMetrics());
     }
